refactor(tui): move app inline tests
Move the current crates/tui/src/tui/app.rs inline test module into crates/tui/src/tui/app/tests.rs. This is a mechanical #3307 extraction and does not change production logic or assertions.

Verification:\n- cargo fmt --all -- --check\n- git diff --check\n- cargo test -p codewhale-tui --bin codewhale-tui --locked tui::app::tests
